redesign the backend + testing GSorter function
