+++ server-side/services/GSorter/govProcessor.js
@@ -1,0 +1,3 @@
+//Generic function to calculate, sort data for summary table
+
+export summaryTable =
