+++ server-side/utils/test.js
@@ -1,0 +1,2 @@
+const { keywordCategoryMap } = require('./categoryHelper');
+console.log(keywordCategoryMap);
